Implement read_attribute() for onoff clsuter.
With this we can read from our mutexed updateState data - now users
can receive updates to state just by taking a ref to updateState and
mutating it. Our reads will pick it up.

This feels hacky - and it has to be redone fo every single type of cluster
as the structure will depend on the clusters attributes - but there
definitely should be a better way to generalize this as opt-in for all
application clusters.